+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## [Unreleased]
+## [2.1.0] - 2024-01-27
 
 ### Added